Add action to woocommerce class
