+++ plugins/woocommerce/src/Admin/Features/OnboardingTasks/Tasks/LaunchYourStore.php
@@ -10,15 +10,6 @@ use Automattic\WooCommerce\Admin\WCAdminHelper;
  * Launch Your Store Task
  */
 class LaunchYourStore extends Task {
-
-	/**
-	 * Constructor.
-	 */
-	public function __construct() {
-		add_action( 'woocommerce_installed', array( $this, 'add_default_values' ) );
-		add_action( 'woocommerce_updated', array( $this, 'add_default_values' ) );
-	}
-
 	/**
 	 * ID.
 	 *
@@ -82,28 +73,5 @@ class LaunchYourStore extends Task {
 	 */
 	public function can_view() {
 		return Features::is_enabled( 'launch-your-store' );
-	}
-
-	/**
-	 * Set default option values for the task.
-	 */
-	public function add_default_values() {
-		$coming_soon      = current_action() === 'woocommerce_installed' ? 'yes' : 'no';
-		$store_pages_only = WCAdminHelper::is_site_fresh() ? 'yes' : 'no';
-		$private_link     = 'yes';
-		$share_key        = wp_generate_password( 32, false );
-
-		if ( ! get_option( 'woocommerce_coming_soon' ) ) {
-			update_option( 'woocommerce_coming_soon', $coming_soon );
-		}
-		if ( ! get_option( 'woocommerce_store_pages_only' ) ) {
-			update_option( 'woocommerce_store_pages_only', $store_pages_only );
-		}
-		if ( ! get_option( 'woocommerce_private_link' ) ) {
-			update_option( 'woocommerce_private_link', $private_link );
-		}
-		if ( ! get_option( 'woocommerce_share_key' ) ) {
-			update_option( 'woocommerce_share_key', $share_key );
-		}
 	}
 }
